Added mentions.test.js file
