+++ app/mentions/mentions.js
@@ -1,14 +1,33 @@
+'use strict';
+
 var express = require('express');
-//var router  = require('router');
+// var router  = require('router');
 //var path    = require('path');
 
-var postObj = {
-  text: 'hola',
-  mention: 'LDougher',
-  username: 'buddy',
-  date: 'Thu Aug 20 2015 14:34:20 GMT-0500 (CDT)',
-  geolocation: 'nashville',
-  hidden: false
-};
+var _ 			= require('lodash');
 
-module.exports = postObj;
+function Mentions(m) {
+	this.username = m.username;
+}
+
+Object.defineProperty(Mentions, 'collection', {
+	get: function () {
+		return mongo.getDb().collection('posts');
+	}
+});
+
+
+
+
+// var postObj = {
+//   text: 'hola',
+//   mention: 'LDougher',
+//   username: 'buddy',
+//   date: 'Thu Aug 20 2015 14:34:20 GMT-0500 (CDT)',
+//   geolocation: 'nashville',
+//   hidden: false
+// };
+
+// module.exports = postObj;
+
+module.exports 		= Mentions;
