+++ app/mentions/routes.js
@@ -3,6 +3,7 @@
 var express = require('express');
 var router = express.Router();
 
+var test  = require('./mentions.test');
 var tests = require('./test/unit/unit');
